2020-07-14预生产环境上线

See merge request mgs/order/order-api!108

+++ src/main/java/com/aiqin/mgs/order/api/service/impl/ReturnOrderInfoServiceImpl.java
@@ -1,8 +1,11 @@
 package com.aiqin.mgs.order.api.service.impl;
 
+import com.aiqin.ground.util.exception.GroundRuntimeException;
 import com.aiqin.ground.util.http.HttpClient;
 import com.aiqin.ground.util.id.IdUtil;
 import com.aiqin.ground.util.json.JsonUtil;
+import com.aiqin.ground.util.protocol.MessageId;
+import com.aiqin.ground.util.protocol.Project;
 import com.aiqin.ground.util.protocol.http.HttpResponse;
 import com.aiqin.mgs.order.api.base.ConstantData;
 import com.aiqin.mgs.order.api.base.PageRequestVO;
@@ -30,12 +33,14 @@ import com.aiqin.mgs.order.api.domain.po.order.ErpOrderOperationLog;
 import com.aiqin.mgs.order.api.domain.request.StoreQuotaRequest;
 import com.aiqin.mgs.order.api.domain.request.returnorder.*;
 import com.aiqin.mgs.order.api.domain.response.returnorder.ReturnOrderStatusVo;
+import com.aiqin.mgs.order.api.domain.response.returnorder.WholesaleReturnList;
 import com.aiqin.mgs.order.api.service.CopartnerAreaService;
 import com.aiqin.mgs.order.api.service.bill.RejectRecordService;
 import com.aiqin.mgs.order.api.service.order.ErpOrderInfoService;
 import com.aiqin.mgs.order.api.service.order.ErpOrderItemService;
 import com.aiqin.mgs.order.api.service.order.ErpOrderQueryService;
 import com.aiqin.mgs.order.api.service.returnorder.ReturnOrderInfoService;
+import com.aiqin.mgs.order.api.util.ResultModel;
 import com.aiqin.platform.flows.client.constant.AjaxJson;
 import com.aiqin.platform.flows.client.constant.FormUpdateUrlType;
 import com.aiqin.platform.flows.client.constant.StatusEnum;
@@ -58,6 +63,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.Assert;
 
 import javax.annotation.Resource;
+import javax.validation.constraints.NotBlank;
 import java.math.BigDecimal;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -85,6 +91,10 @@ public class ReturnOrderInfoServiceImpl implements ReturnOrderInfoService {
 
     @Value("${bridge.url.product-api}")
     private String productHost;
+
+    @Value("${bridge.url.scmp-api}")
+    private String scmpHost;
+
 
     @Autowired
     private ReturnOrderInfoDao returnOrderInfoDao;
@@ -131,6 +141,19 @@ public class ReturnOrderInfoServiceImpl implements ReturnOrderInfoService {
         if(aBoolean){
             reallyReturn=1;
         }
+        //获取加盟商以及合伙人
+        ReturnOrderFranchisee returnOrderFranchisee = erpOrderInfoDao.selectFranchisee(reqVo.getOrderStoreCode());
+        if (returnOrderFranchisee != null){
+            reqVo.setFranchiseeCode(returnOrderFranchisee.getFranchiseeCode());
+            reqVo.setFranchiseeName(returnOrderFranchisee.getFranchiseeName());
+            reqVo.setCopartnerAreaId(returnOrderFranchisee.getCopartnerAreaId());
+            reqVo.setCopartnerAreaName(returnOrderFranchisee.getCopartnerAreaName());
+        }
+        //业务形式  0门店形式  1批发形式
+//        String storeCode = reqVo.getStoreCode();
+//        if (!storeCode.equals(null)){
+            reqVo.setBusinessForm(0);
+//        }
         ReturnOrderInfo record = new ReturnOrderInfo();
         Date now = new Date();
         BeanUtils.copyProperties(reqVo, record);
@@ -376,9 +399,25 @@ public class ReturnOrderInfoServiceImpl implements ReturnOrderInfoService {
         }
         if(cancelFlag){
             //通知退供单-撤销
-            log.info("通知退供单-撤销开始，rejectRecordCode={}",reqVo.getReturnOrderCode());
-            rejectRecordService.removeRejectRecordStatus(reqVo.getReturnOrderCode());
-            log.info("通知退供单-撤销结束");
+            //查询退货单是否同步成功
+            Integer integer = returnOrderInfoDao.selectType(reqVo.getReturnOrderCode());
+            if (integer.equals(2)){
+                log.info("通知供应链退货单取消-开始");
+                String returnOrderCode = reqVo.getReturnOrderCode();
+                String url =scmpHost+"/returnGoods/cancel";
+                HttpClient httpClient = HttpClient.get(url).addParameter("return_order_code", returnOrderCode).timeout(30000);
+                AjaxJson result = httpClient.action().result(AjaxJson.class);
+                log.info("通知供应链返回结果：{}",JSON.toJSONString(result));
+                if(result.getSuccess()){
+                   log.info("通知供应链-退货单取消-通知完成");
+                }else {
+                    log.info("通知供应链-退货单取消-通知失败");
+                }
+            }else {
+                log.info("通知退供单-撤销开始，rejectRecordCode={}", reqVo.getReturnOrderCode());
+                rejectRecordService.removeRejectRecordStatus(reqVo.getReturnOrderCode());
+                log.info("通知退供单-撤销结束");
+            }
         }
         //配送且一般退货调用且审核通过，调用门店退货申请-完成(门店)
         if("15".equals(returnOrderInfo.getReturnReasonCode())&&returnOrderInfo.getOrderType().equals(2)&&sysFlag){
@@ -763,10 +802,17 @@ public class ReturnOrderInfoServiceImpl implements ReturnOrderInfoService {
             //1直送 2配送 3货架
             Integer type=returnOrderInfo.getOrderType();
             if(null!=type&&type.equals(1)){//订单类型 14配送tob 2直送tob
-                json.put("pay_order_type", PayOrderTypeEnum.PAY_ORDER_TYPE_PEI.getKey());
-                json.put("pay_origin_type",PayOriginTypeEnum.DIRECT_SEND_TOB_RETURN.getKey());
+//                json.put("pay_order_type", PayOrderTypeEnum.PAY_ORDER_TYPE_PEI.getKey());
+                json.put("pay_order_type", PayOrderTypeEnum.PAY_ORDER_TYPE_ZHI.getKey());
+//                json.put("pay_origin_type",PayOriginTypeEnum.DIRECT_SEND_TOB_RETURN.getKey());
+                json.put("pay_origin_type",PayOriginTypeEnum.TOB_RETURN.getKey());
             }else if(null!=type&&type.equals(2)){
-                json.put("pay_order_type",PayOrderTypeEnum.PAY_ORDER_TYPE_ZHI.getKey());
+//                json.put("pay_order_type",PayOrderTypeEnum.PAY_ORDER_TYPE_ZHI.getKey());
+                json.put("pay_order_type",PayOrderTypeEnum.PAY_ORDER_TYPE_PEI.getKey());
+//                json.put("pay_origin_type",PayOriginTypeEnum.TOB_RETURN.getKey());
+                json.put("pay_origin_type",PayOriginTypeEnum.DIRECT_SEND_TOB_RETURN.getKey());
+            }else if (null!=type&&type.equals(3)){
+                json.put("pay_order_type", PayOrderTypeEnum.PAY_ORDER_TYPE_ZHI.getKey());
                 json.put("pay_origin_type",PayOriginTypeEnum.TOB_RETURN.getKey());
             }
             json.put("back_url","http://order.api.aiqin.com/returnOrder/callback");
@@ -829,11 +875,7 @@ public class ReturnOrderInfoServiceImpl implements ReturnOrderInfoService {
         List<ReturnOrderDetail> returnOrderDetails = returnOrderDetailDao.selectListByReturnOrderCode(returnOrderCode);
         //退货总金额(赠品金额+商品金额已算好) 加
         BigDecimal returnOrderAmount = returnOrderInfo.getReturnOrderAmount();
-        BigDecimal topCouponDiscountAmount =returnOrderInfo.getTopCouponDiscountAmount();
-        if(null==topCouponDiscountAmount){
-            topCouponDiscountAmount=BigDecimal.ZERO;
-        }
-        returnOrderInfo.setReturnOrderAmount(returnOrderAmount.add(topCouponDiscountAmount));
+        returnOrderInfo.setReturnOrderAmount(returnOrderAmount.add(returnOrderInfo.getTopCouponDiscountAmount() == null ? BigDecimal.ZERO : returnOrderInfo.getTopCouponDiscountAmount()));
 //        BigDecimal reduce = returnOrderDetails.stream().map(ReturnOrderDetail::getTopCouponDiscountAmount).reduce(BigDecimal.ZERO, BigDecimal::add);
         //查询日志详情
         List<ErpOrderOperationLog> erpOrderOperationLogs = getOrderOperationLogList(returnOrderCode);
@@ -876,7 +918,8 @@ public class ReturnOrderInfoServiceImpl implements ReturnOrderInfoService {
         //保存数据权限查询出的数据
         List<String> storesIds=new ArrayList<>();
         List<String> newIds=new ArrayList<>();
-        if(searchVo.getSearchVO().getOrderType().equals(2)&&"15".equals(searchVo.getSearchVO().getReturnReasonCode())){//一般退货需要加数据权限
+//        if(searchVo.getSearchVO().getOrderType().equals(2)&&"15".equals(searchVo.getSearchVO().getReturnReasonCode())){//一般退货需要加数据权限
+        if(searchVo.getSearchVO().getOrderType().equals(1)&&"15".equals(searchVo.getSearchVO().getReturnReasonCode())){//一般退货需要加数据权限
             if(StringUtils.isBlank(searchVo.getSearchVO().getPersonId())||StringUtils.isBlank(searchVo.getSearchVO().getResourceCode())){
                 return pageResData;
             }
@@ -984,7 +1027,7 @@ public class ReturnOrderInfoServiceImpl implements ReturnOrderInfoService {
     public HttpResponse saveWriteDownOrder(String orderCode) {
         log.info("发起冲减单开始,原始订单编码,orderCode={}",orderCode);
         //根据订单编码查询原始订单数据及详情数据
-        ErpOrderInfo erpOrderInfo=erpOrderQueryService.getOrderAndItemByOrderCode(orderCode);
+        ErpOrderInfo erpOrderInfo= erpOrderQueryService.getOrderAndItemByOrderCode(orderCode);
         if(null==erpOrderInfo){
             //此单号有误，未查到订单数据
             return HttpResponse.failure(ResultCode.NOT_FOUND_ORDER_DATA);
@@ -1049,7 +1092,6 @@ public class ReturnOrderInfoServiceImpl implements ReturnOrderInfoService {
                     returnOrderDetail.setBarCode(eoi.getBarCode());
                     //批次
                     returnOrderDetail.setBatchInfoCode(eoi.getBatchInfoCode());
-//                    returnOrderDetail.setBatchNumber(eoi.getBatchNumber());
                     returnOrderDetail.setBatchCode(eoi.getBatchCode());
                     returnOrderDetail.setBatchDate(eoi.getBatchDate());
                     //可用赠品额度
@@ -1189,7 +1231,7 @@ public class ReturnOrderInfoServiceImpl implements ReturnOrderInfoService {
        Map<String, Object> result = null;
        result = httpClient.action().result(new TypeReference<Map<String, Object>>() {
        });
-       log.info("发起退积分申请结果，request={}",JsonUtil.toJson(result));
+       log.info("发起退积分申请结果，request={}",JSON.toJSON(result));
        if(result!=null&&"0".equals(result.get("code"))){
            log.info("退积分完成");
            return true;
@@ -1596,6 +1638,170 @@ public class ReturnOrderInfoServiceImpl implements ReturnOrderInfoService {
         Integer pageCount = returnOrderInfoDao.pageCount(searchVo);
         return new PageResData<>(pageCount, content);
     }
+
+    /**
+     * 批发退货
+     * @param reqVo
+     * @return
+     */
+    @Override
+    @Transactional(rollbackFor = Exception.class)
+    public HttpResponse saveWholesaleReturn(ReturnWholesaleOrderReqVo reqVo) {
+        log.info("发起批发退货--入参"+JSON.toJSONString(reqVo));
+        try {
+            //查询订单是否存在未处理售后单
+            List<ReturnOrderInfo> returnOrderInfo = returnOrderInfoDao.selectByOrderCodeAndStatus(reqVo.getOrderStoreCode(), 1);
+            Assert.isTrue(CollectionUtils.isEmpty(returnOrderInfo), "该订单还有未审核售后单，请稍后提交");
+            //校验原始订单的主订单关联的所有子订单是否发货完成
+            ErpOrderInfo orderByOrderCode = erpOrderQueryService.getOrderByOrderCode(reqVo.getOrderStoreCode());
+            Boolean aBoolean = checkSendOk(orderByOrderCode.getMainOrderCode());
+            //是否真的发起退货 0:预生成退货单 1:原始订单全部发货完成生成退货单
+            Integer reallyReturn = 0;
+            if (aBoolean) {
+                reallyReturn = 1;
+            }
+            //售后类型
+            reqVo.setReturnOrderType(2);
+            //加盟商和合伙人
+            ReturnOrderFranchisee returnOrderFranchisee = erpOrderInfoDao.selectFranchisee(reqVo.getOrderStoreCode());
+            if (returnOrderFranchisee != null) {
+                reqVo.setFranchiseeCode(returnOrderFranchisee.getFranchiseeCode());
+                reqVo.setFranchiseeName(returnOrderFranchisee.getFranchiseeName());
+                reqVo.setCopartnerAreaId(returnOrderFranchisee.getCopartnerAreaId());
+                reqVo.setCopartnerAreaName(returnOrderFranchisee.getCopartnerAreaName());
+            }
+            ReturnOrderInfo record = new ReturnOrderInfo();
+            Date now = new Date();
+            BeanUtils.copyProperties(reqVo, record);
+            String returnOrderId = IdUtil.uuid();
+            String afterSaleCode = sequenceService.generateOrderAfterSaleCode(reqVo.getCompanyCode(), reqVo.getReturnOrderType());
+            record.setReturnOrderId(returnOrderId);
+            record.setReturnOrderCode(afterSaleCode);
+            record.setCreateTime(now);
+            //判断是ERP还是批发客户发起的退货
+            if (reqVo.getTreatmentMethod().equals(1) && reqVo.getOrderCategory().equals(51)) {
+                record.setReturnOrderStatus(ReturnOrderStatusEnum.RETURN_ORDER_STATUS_COM.getKey());
+                //退货类型
+                record.setReturnOrderType(2);
+                //对应货架退货补货单
+            } else if (reqVo.getTreatmentMethod().equals(1) && reqVo.getOrderCategory().equals(17)) {
+                record.setReturnOrderStatus(ReturnOrderStatusEnum.RETURN_ORDER_STATUS_COM.getKey());
+                //退货类型
+                record.setReturnOrderType(reqVo.getReturnOrderTypeTemporary());
+                //对应普通补货退货单
+            } else if (reqVo.getTreatmentMethod().equals(1) && reqVo.getOrderCategory().equals(147)) {
+                record.setReturnOrderStatus(ReturnOrderStatusEnum.RETURN_ORDER_STATUS_COM.getKey());
+                record.setReturnOrderType(reqVo.getReturnOrderTypeTemporary());
+            } else {
+                //批发客户发起的退货状态为审核中
+                record.setReturnOrderStatus(ReturnOrderStatusEnum.RETURN_ORDER_STATUS_WAIT.getKey());
+            }
+            //实退商品数量
+            record.setActualProductCount(record.getProductCount());
+            //实退商品总金额
+            record.setActualReturnOrderAmount(record.getReturnOrderAmount());
+//            //退货类型
+//            record.setReturnOrderType(2);
+            //退货单--退款方式 1:现金 2:微信 3:支付宝 4:银联 5:退到加盟商账户
+            record.setReturnMoneyType(ConstantData.RETURN_MONEY_TYPE);
+            record.setOrderSuccess(1);
+            record.setReallyReturn(reallyReturn);
+            //退货金额
+            BigDecimal returnOrderAmount = BigDecimal.ZERO;
+            //普通循环
+            List<ReturnOrderDetail> details = new ArrayList<>();
+            List<ReturnWholesaleOrderDetail> details1 = reqVo.getDetails();
+            for (ReturnWholesaleOrderDetail sd : details1) {
+                ReturnOrderDetail detail = new ReturnOrderDetail();
+                    //商品属性 0新品1残品
+                    Integer productStatus = 0;
+                    if (null != sd.getProductStatus()) {
+                        productStatus = sd.getProductStatus();
+                    }
+                    //均摊后的金额乘以退货数量
+                    BigDecimal preferentialAmount = sd.getPreferentialAmount();
+                    BigDecimal multiply = preferentialAmount.multiply(BigDecimal.valueOf(sd.getReturnProductCount()));
+                    returnOrderAmount = returnOrderAmount.add(multiply);
+                    BeanUtils.copyProperties(sd, detail);
+                    detail.setCreateTime(now);
+                    detail.setReturnOrderDetailId(IdUtil.uuid());
+                    detail.setReturnOrderCode(afterSaleCode);
+                    detail.setCreateById(reqVo.getCreateById());
+                    detail.setCreateByName(reqVo.getCreateByName());
+                    detail.setProductStatus(productStatus);
+                    details.add(detail);
+                }
+            //退货金额
+            record.setReturnOrderAmount(returnOrderAmount);
+            log.info("发起批发退货--插入批发退货信息，record={}", record);
+            returnOrderInfoDao.insertSelective(record);
+            log.info("发起批发退货--插入批发退货详情，details={}", details);
+            returnOrderDetailDao.insertBatch(details);
+            //添加日志
+            log.info("发起批发退货--插入日志，details={}",details);
+            insertLog(afterSaleCode,reqVo.getCreateById(),reqVo.getCreateByName(),ErpLogOperationTypeEnum.ADD.getCode(),ErpLogSourceTypeEnum.RETURN.getCode(),
+                    record.getReturnOrderStatus().equals(ReturnOrderStatusEnum.RETURN_ORDER_STATUS_WAIT.getKey()) ? record.getReturnOrderStatus() : ReturnOrderStatusEnum.RETURN_ORDER_STATUS_COM.getKey(),
+                    record.getReturnOrderStatus().equals(ReturnOrderStatusEnum.RETURN_ORDER_STATUS_WAIT.getKey()) ? ReturnOrderStatusEnum.RETURN_ORDER_STATUS_WAIT.getMsg() : ReturnOrderStatusEnum.RETURN_ORDER_STATUS_COM.getMsg());
+            //修改原始订单数据
+            log.info("发起批发退货--修改原始订单数据开始,入参orderStoreCode={},orderReturnStatusEnum={},returnQuantityList={},personId={},personName={}", record.getOrderStoreCode(), ErpOrderReturnStatusEnum.WAIT, null, record.getCreateById(), record.getCreateByName());
+            erpOrderInfoService.updateOrderReturnStatus(record.getOrderStoreCode(), ErpOrderReturnRequestEnum.WAIT, null, record.getCreateById(), record.getCreateByName());
+            log.info("发起批发退货--修改原始订单数据结束");
+            log.info("审核后-调用发起批发退货开始");
+            ReturnOrderReviewReqVo reqVo1 = new ReturnOrderReviewReqVo();
+            reqVo1.setOperateStatus(reqVo.getTreatmentMethod());
+            reqVo1.setReturnOrderCode(afterSaleCode);
+            log.info("erp-批发退货-同步供应链-生成退供单开始---");
+            if (reallyReturn == 0) {  //预生成退货单不同步供应链
+            }else {
+                HttpResponse httpResponse = updateReturnStatus(reqVo1);
+//             updateReturnStatus(reqVo1);
+                log.info("erp-批发退货-同步供应链，生成退供单结束,httpResponse={}", JSON.toJSON(httpResponse));
+//            log.info("erp-批发退货-同步供应链，生成退供单结束");
+                if (!"0".equals(httpResponse.getCode())) {
+                    //erp同步供应链，生成退供单失败
+                    throw new RuntimeException("erp-批发退货-同步供应链，生成退供单失败");
+                }
+                log.info("erp-批发退货-同步供应链，生成退货单成功");
+              }
+                 return HttpResponse.success();
+            }catch(Exception e){
+                log.error("批发退货-请求：{},{}", reqVo, e);
+                throw new GroundRuntimeException("发起批发退货出现未知异常,请联系系统管理员");
+            }
+
+
+    }
+
+    /**
+     * 多条件查询批发退货列表
+     * @param whoVo
+     * @return
+     */
+    @Override
+    public HttpResponse selectAllList(wholesaleReturnOrderSearchVo whoVo) {
+        log.info("多条件查询批发退货列表-入参：{}",whoVo);
+        Integer PageNo = whoVo.getPageNo();
+        Integer PageSize = whoVo.getPageSize();
+        if (PageNo == null && PageSize == null) {
+            PageNo = 1;
+            PageSize = 10;
+        }
+        if (PageNo == 0 && PageSize == 0) {
+            PageNo = 1;
+            PageSize = 10;
+        }
+        PageHelper.startPage(PageNo, PageSize);
+        List<WholesaleReturnList> wholesaleReturnLists =  returnOrderInfoDao.selectByCondition(whoVo);
+        log.info("条件分页查询返回参数对象集合:{}", wholesaleReturnLists);
+        if (wholesaleReturnLists != null){
+            ResultModel resultModel = new ResultModel();
+            resultModel.setResult(wholesaleReturnLists);
+            resultModel.setTotal(((Page) wholesaleReturnLists).getTotal());
+            return HttpResponse.success(resultModel);
+        }
+        return HttpResponse.success();
+    }
+
 
     /**
      * 插入日志表
